requirements.txt
Flask==2.0.1
selenium==3.141.0
urllib3==1.26.5
beautifulsoup4==4.9.3
cloudscraper==1.2.58
requests==2.25.1
python-dotenv==0.19.1
user_agent==0.1.9
